// libs/modules/global/secrets/enum.ts
export enum AUTH_SERVICE_ENVIRONMENT {
  PORT = 'AUTH_SERVICE_PORT',
  URL = 'AUTH_SERVICE_URL',
}

export enum NOTIFICATION_SERVICE_ENVIRONMENT {
  PORT = 'NOTIFICATION_SERVICE_PORT',
}

export enum DEFAULT_DB_ENVIRONMENT {
  HOST = 'DEFAULT_DB_HOST',
  PORT = 'DEFAULT_DB_PORT',
  USERNAME = 'DEFAULT_DB_USERNAME',
  PASSWORD = 'DEFAULT_DB_PASSWORD',
  DATABASE = 'DEFAULT_DB_DATABASE',
  SLAVES = 'DEFAULT_DB_SLAVES',
}